third_party/WebKit/LayoutTests/fast/css/pseudo-valid-dynamic-expected.txt

This test checks that input element that share style, no longer share style after validity rules are dynamically added.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 
PASS document.defaultView.getComputedStyle(v[0], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[1], null).getPropertyValue('background-color') is 'rgb(0, 0, 255)'
PASS successfullyParsed is true

TEST COMPLETE
